requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-ckeditor==6.1.0
django-colorfield==0.4.3
django-cors-headers==3.8.0
django-easy-select2==1.5.7
django-extensions==3.1.3
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-mptt==0.13.3
django-nested-inline==0.4.4
django-ninja==0.15.0
django-silk==4.1.0
django-taggit==1.5.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
gprof2dot==2021.2.21
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pydantic==1.8.2
Pygments==2.10.0
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
